refactor: Move error-event.ts
