+++ packages/server/src/error-event.ts
