+++ packages/server/src/internals/swr-queue.ts
@@ -5,12 +5,12 @@ import * as R from 'ramda'
 
 import { createAuthServices } from './authentication'
 import { isLegacyQuery, LegacyQuery } from './data-source-helper'
-import { captureErrorEvent } from './error-event'
 import { log } from './log'
 import { type Context } from '~/context'
 import { CacheEntry, Cache, Priority } from '~/context/cache'
 import { SwrQueue } from '~/context/swr-queue'
 import { Database } from '~/database'
+import { captureErrorEvent } from '~/error-event'
 import { modelFactories } from '~/model'
 import { cachedResolvers } from '~/schema'
 import { Timer, Time, timeToSeconds, timeToMilliseconds } from '~/timer'
